6.1.2 Перенесет код отрисовки точки маршрута из trip-controller в point-controller
